requirements.txt
crayons
discord.py